+++ example/rails-32/config/initializers/bugsnag.rb
@@ -1,3 +1,0 @@
-Bugsnag.configure do |config|
-  config.api_key = "YOUR_API_KEY_HERE"
-end