+++ features/fixtures/plain/app/configuration/api_key.rb
@@ -1,0 +1,9 @@
+require './app'
+
+configure_basics
+
+Bugsnag.configure do |conf|
+  conf.api_key = ENV["MAZE_API_KEY"]
+end
+
+Bugsnag.notify(RuntimeError.new("Oh no"))