+++ features/fixtures/plain/app/exception_data/handled_hash.rb
@@ -1,0 +1,10 @@
+require './app'
+require './exception_data/crash'
+
+configure_basics
+
+begin
+  exception_with_hash
+rescue => exception
+  Bugsnag.notify(exception)
+end