+++ features/fixtures/plain/app/stack_frame_modification/initiators/handled_before_notify.rb
@@ -1,0 +1,25 @@
+require 'bugsnag'
+require './app'
+
+configure_basics
+
+def run(callback)
+  Bugsnag.before_notify_callbacks << callback
+  step_one
+end
+
+def step_one
+  step_two
+end
+
+def step_two
+  step_three
+end
+
+def step_three
+  crash
+end
+
+def crash
+  Bugsnag.notify(RuntimeError.new("oh no"))
+end