+++ features/fixtures/rails4/app/app/controllers/unhandled_controller.rb
@@ -1,0 +1,11 @@
+class UnhandledController < ActionController::Base
+  protect_from_forgery with: :exception
+
+  def index
+    render json: {}
+  end
+
+  def error
+    generate_unhandled_error
+  end
+end