+++ spec/bugsnag_spec.rb
@@ -17,6 +17,62 @@ describe Bugsnag do
     end
   end
 
+  describe "add_exit_handler" do
+
+    before do
+      Bugsnag.instance_variable_set(:@exit_handler_added, false)
+    end
+
+    it "automatically adds an exit handler" do
+      expect(Bugsnag).to receive(:register_at_exit)
+      Bugsnag.configure do |conf|
+        conf.api_key = "TEST KEY"
+      end
+    end
+
+    it "calls at_exit when register_at_exit is called" do
+      expect(Bugsnag).to receive(:at_exit)
+      Bugsnag.register_at_exit
+    end
+
+    it "doesn't call at_exit on subsequent calls" do
+      expect(Bugsnag).to receive(:at_exit).once
+      Bugsnag.register_at_exit
+      Bugsnag.register_at_exit
+    end
+
+    context 'with aliased at_exit' do
+      before do
+        module Kernel
+          alias_method :old_at_exit, :at_exit
+          def at_exit
+            begin
+              raise BugsnagTestException.new("Oh no")
+            rescue
+              yield
+            end
+          end
+        end
+      end
+
+      it "sends an exception when at_exit is called" do
+        report_mock = double('report')
+        expect(report_mock).to receive(:severity=).with('error')
+        expect(report_mock).to receive(:severity_reason=).with({
+          :type => Bugsnag::Report::UNHANDLED_EXCEPTION
+        })
+        expect(Bugsnag).to receive(:notify).with(kind_of(BugsnagTestException), true).and_yield(report_mock)
+        Bugsnag.register_at_exit
+      end
+
+      after do
+        module Kernel
+          alias_method :at_exit, :old_at_exit
+        end
+      end
+    end
+  end
+
   describe 'loading integrations' do
     before do
       module Kernel
@@ -33,7 +89,6 @@ describe Bugsnag do
     end
 
     it 'attempts to load integrations' do
-      Kernel::REQUIRED = []
       ENV["BUGSNAG_DISABLE_AUTOCONFIGURE"] = nil
       load "./lib/bugsnag.rb"
       Bugsnag::INTEGRATIONS.each do |integration|
@@ -42,36 +97,31 @@ describe Bugsnag do
     end
 
     it 'does not load integrations when BUGSNAG_DISABLE_AUTOCONFIGURE is true' do
-      Kernel::REQUIRED = []
       ENV["BUGSNAG_DISABLE_AUTOCONFIGURE"] = 'true'
       load "./lib/bugsnag.rb"
       expect(Kernel::REQUIRED).to eq(["bugsnag/integrations/rack"])
     end
 
     it 'loads all integrations if requested' do
-      Kernel::REQUIRED = []
-      expect(Kernel::REQUIRED).to eq([])
       Bugsnag.load_integrations
       Bugsnag::INTEGRATIONS.each do |integration|
         expect(Kernel::REQUIRED).to include("bugsnag/integrations/#{integration}")
       end
     end
 
-    it 'loads singular integrations' do
-      Kernel::REQUIRED = []
-      expect(Kernel::REQUIRED).to eq([])
-      Bugsnag::INTEGRATIONS.each do |integration|
-        Kernel::REQUIRED = []
+    Bugsnag::INTEGRATIONS.each do |integration|
+      it "loads #{integration}" do
         Bugsnag.load_integration(integration)
         expect(Kernel::REQUIRED).to include("bugsnag/integrations/#{integration}")
       end
     end
 
-    it 'loads railtie for :rails or :railtie' do
-      Kernel::REQUIRED = []
+    it 'loads railtie for rails' do
       Bugsnag.load_integration(:rails)
       expect(Kernel::REQUIRED).to include("bugsnag/integrations/railtie")
-      Kernel::REQUIRED = []
+    end
+
+    it 'loads railtie for railtie' do
       Bugsnag.load_integration(:railtie)
       expect(Kernel::REQUIRED).to include("bugsnag/integrations/railtie")
     end