+++ spec/configuration_spec.rb
@@ -29,7 +29,197 @@ describe Bugsnag::Configuration do
     end
   end
 
+  describe "set_proxy" do
+    it "defaults proxy settings to nil" do
+      expect(subject.proxy_host).to be nil
+      expect(subject.proxy_port).to be nil
+      expect(subject.proxy_user).to be nil
+      expect(subject.proxy_password).to be nil
+    end
+
+    it "allows proxy settings to be set directly" do
+      subject.proxy_host = "http://localhost"
+      subject.proxy_port = 34000
+      subject.proxy_user = "user"
+      subject.proxy_password = "password"
+      expect(subject.proxy_host).to eq("http://localhost")
+      expect(subject.proxy_port).to eq(34000)
+      expect(subject.proxy_user).to eq("user")
+      expect(subject.proxy_password).to eq("password")
+    end
+
+    it "parses a uri if provided" do
+      subject.parse_proxy("http://user:password@localhost:34000")
+      expect(subject.proxy_host).to eq("localhost")
+      expect(subject.proxy_port).to eq(34000)
+      expect(subject.proxy_user).to eq("user")
+      expect(subject.proxy_password).to eq("password")
+    end
+
+    it "automatically parses http_proxy environment variable" do
+      ENV['http_proxy'] = "http://user:password@localhost:34000"
+      expect(subject.proxy_host).to eq("localhost")
+      expect(subject.proxy_port).to eq(34000)
+      expect(subject.proxy_user).to eq("user")
+      expect(subject.proxy_password).to eq("password")
+    end
+
+    it "automatically parses https_proxy environment variable" do
+      ENV['https_proxy'] = "https://user:password@localhost:34000"
+      expect(subject.proxy_host).to eq("localhost")
+      expect(subject.proxy_port).to eq(34000)
+      expect(subject.proxy_user).to eq("user")
+      expect(subject.proxy_password).to eq("password")
+    end
+
+    after do
+      ENV['http_proxy'] = nil
+      ENV['https_proxy'] = nil
+    end
+  end
+
+  describe "logger" do
+    class TestLogger
+      attr_accessor :logs
+
+      def initialize
+        @logs = []
+      end
+
+      def log(level, name, &block)
+        message = block.call
+        @logs << {
+          :level => level,
+          :name => name,
+          :message => message
+        }
+      end
+
+      def info(name, &block)
+        log('info', name, &block)
+      end
+
+      def warn(name, &block)
+        log('warning', name, &block)
+      end
+
+      def debug(name, &block)
+        log('debug', name, &block)
+      end
+    end
+
+    before do
+      @logger = TestLogger.new
+      Bugsnag.configure do |bugsnag|
+        bugsnag.logger = @logger
+      end
+    end
+
+    context "using configure" do
+      before do
+        Bugsnag.configuration.api_key = nil
+        Bugsnag.instance_variable_set("@key_warning", nil)
+        ENV['BUGSNAG_API_KEY'] = nil
+        expect(@logger.logs.size).to eq(0)
+      end
+
+      context "API key is not specified" do
+        it "skips logging a warning if validate_api_key is false" do
+          Bugsnag.configure(false)
+          expect(@logger.logs.size).to eq(0)
+        end
+
+        it "logs a warning by default" do
+          Bugsnag.configure
+          expect(@logger.logs.size).to eq(1)
+          log = @logger.logs.first
+          expect(log).to eq({
+            :level => "warning",
+            :name => "[Bugsnag]",
+            :message => "No valid API key has been set, notifications will not be sent"
+          })
+        end
+
+        it "logs a warning if validate_api_key is true" do
+          Bugsnag.configure(true)
+          expect(@logger.logs.size).to eq(1)
+          log = @logger.logs.first
+          expect(log).to eq({
+            :level => "warning",
+            :name => "[Bugsnag]",
+            :message => "No valid API key has been set, notifications will not be sent"
+          })
+        end
+      end
+
+      context "API key is set" do
+        it "skips logging a warning when configuring with an API key" do
+          Bugsnag.configure do |config|
+            config.api_key = 'd57a2472bd130ac0ab0f52715bbdc600'
+          end
+          expect(@logger.logs.size).to eq(0)
+        end
+
+        it "logs a warning if the configured API key is invalid" do
+          Bugsnag.configure do |config|
+            config.api_key = 'WARNING: not a real key'
+          end
+          expect(@logger.logs.size).to eq(1)
+          log = @logger.logs.first
+          expect(log).to eq({
+            :level => "warning",
+            :name => "[Bugsnag]",
+            :message => "No valid API key has been set, notifications will not be sent"
+          })
+        end
+      end
+    end
+
+    it "should log info messages to the set logger" do
+      expect(@logger.logs.size).to eq(0)
+      Bugsnag.configuration.info("Info message")
+      expect(@logger.logs.size).to eq(1)
+      log = @logger.logs.first
+      expect(log).to eq({
+        :level => "info",
+        :name => "[Bugsnag]",
+        :message => "Info message"
+      })
+    end
+
+    it "should log warning messages to the set logger" do
+      expect(@logger.logs.size).to eq(0)
+      Bugsnag.configuration.warn("Warning message")
+      expect(@logger.logs.size).to eq(1)
+      log = @logger.logs.first
+      expect(log).to eq({
+        :level => "warning",
+        :name => "[Bugsnag]",
+        :message => "Warning message"
+      })
+    end
+
+    it "should log debug messages to the set logger" do
+      expect(@logger.logs.size).to eq(0)
+      Bugsnag.configuration.debug("Debug message")
+      expect(@logger.logs.size).to eq(1)
+      log = @logger.logs.first
+      expect(log).to eq({
+        :level => "debug",
+        :name => "[Bugsnag]",
+        :message => "Debug message"
+      })
+    end
+
+    after do
+      Bugsnag.configure do |bugsnag|
+        bugsnag.logger = Logger.new(StringIO.new)
+      end
+    end
+  end
+
   it "should have exit exception classes ignored by default" do
     expect(subject.ignore_classes).to eq(Set.new([SystemExit, Interrupt]))
   end
+
 end