+++ spec/integrations/rack_spec.rb
@@ -62,11 +62,72 @@ describe Bugsnag::Rack do
     before do
       unless defined?(::Rack)
         @mocked_rack = true
-        class ::Rack
-          class ::Request
+        class Rack
+          class Request
           end
         end
       end
+    end
+
+    it "correctly redacts from url and referer any value indicated by meta_data_filters" do
+      callback = double
+      rack_env = {
+        :env => true,
+        :HTTP_REFERER => "https://bugsnag.com/about?email=hello@world.com&another_param=thing",
+        "rack.session" => {
+          :session => true
+        }
+      }
+
+      rack_request = double
+      rack_params = {
+        :param => 'test'
+      }
+      allow(rack_request).to receive_messages(
+        :params => rack_params,
+        :ip => "rack_ip",
+        :request_method => "TEST",
+        :path => "/TEST_PATH",
+        :scheme => "http",
+        :host => "test_host",
+        :port => 80,
+        :referer => "https://bugsnag.com/about?email=hello@world.com&another_param=thing",
+        :fullpath => "/TEST_PATH?email=hello@world.com&another_param=thing"
+      )
+      expect(::Rack::Request).to receive(:new).with(rack_env).and_return(rack_request)
+
+      # modify rack_env to include redacted referer
+      report = double("Bugsnag::Report")
+      allow(report).to receive(:request_data).and_return({
+        :rack_env => rack_env
+      })
+      expect(report).to receive(:context=).with("TEST /TEST_PATH")
+      expect(report).to receive(:user).and_return({})
+
+      config = double
+      allow(config).to receive(:send_environment).and_return(true)
+      allow(config).to receive(:meta_data_filters).and_return(['email'])
+      allow(report).to receive(:configuration).and_return(config)
+      expect(report).to receive(:add_tab).once.with(:request, {
+        :url => "http://test_host/TEST_PATH?email=[FILTERED]&another_param=thing",
+        :httpMethod => "TEST",
+        :params => rack_params,
+        :referer => "https://bugsnag.com/about?email=[FILTERED]&another_param=thing",
+        :clientIp => "rack_ip",
+        :headers => {
+          "Referer" => "https://bugsnag.com/about?email=[FILTERED]&another_param=thing"
+        }
+      })
+      # rack_env["HTTP_REFERER"] = "https://bugsnag.com/about?email=[FILTERED]&another_param=thing"
+      expect(report).to receive(:add_tab).once.with(:environment, rack_env)
+      expect(report).to receive(:add_tab).once.with(:session, {
+        :session => true
+      })
+
+      expect(callback).to receive(:call).with(report)
+
+      middleware = Bugsnag::Middleware::RackRequest.new(callback)
+      middleware.call(report)
     end
 
     it "correctly extracts data from rack middleware" do
@@ -78,7 +139,7 @@ describe Bugsnag::Rack do
           :session => true
         }
       }
-      
+
       rack_request = double
       rack_params = {
         :param => 'test'
@@ -94,7 +155,7 @@ describe Bugsnag::Rack do
         :referer => "referer",
         :fullpath => "/TEST_PATH"
       )
-      expect(::Rack::Request).to receive(:new).with(rack_env).and_return(rack_request)
+      expect(Rack::Request).to receive(:new).with(rack_env).and_return(rack_request)
 
       report = double("Bugsnag::Report")
       allow(report).to receive(:request_data).and_return({